refactor: fen & pureCoordinateNotation to new subfolder
