+++ src/chess/serialization/fen/fen.constants.ts
