+++ src/chess/serialization/fen/fen.ts
@@ -1,4 +1,4 @@
-import { State } from '../chess.models'
+import { State } from '../../chess.models'
 import {
 	boardToFen,
 	castlingAbilityToFen,
